embedding_service/requirements.txt
# Configuration
python-dotenv==1.0.1
pydantic==1.10.7

# Asynchronous Libraries
aio-pika==9.5.3
aiohttp==3.11.10

# Elasticsearch
elasticsearch[async]==8.8.1

# Image Processing and Embedding
clip @ git+https://github.com/openai/CLIP.git
torchvision==0.20.1
Pillow==9.4.0

# Metrics
prometheus_client==0.16.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1

# FastAPI
fastapi==0.115.2
uvicorn==0.30.6
